03-airflow/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pyspark
pandas
openpyxl
apache-airflow-providers-apache-spark